// src/app/models/solicitud-inspeccion.ts
export interface SolicitudInspeccion {
    nombre: string;
    apellido: string;
    telefono: string;
    email: string;
    barrio: string;
    sector: string;
    manzana: string;
    lote: string;
    fechaSolicitud: Date;
    estado: string;
    numeroExpediente: string;
    tipo_Solicitud: string;
}